initial commit of automatic pseudonymisation for patient data.
when patient records are saved, the patient identifiable fields are passed through a 'randomisedata' function.

this behaviour is enabled by default, but can be disabled through setting a variable in params.php, which i have added to params.sample.php

this code will likely require refactoring or moving out to a service at a later stage.