Adding code to double the time of procedures if you're operating on both eyes (OE-174)
Edge case: If you have both eyes selected, add some procedures, and
then change the eye to be left or right, it will not halve the
procedure times, so they will be off at that point. Likewise if you
delete procedures after changing eyes - it won't subtract the doubled
time, so the total durations will be off.