added a firm_id field to the referral with a foreign key constraint to the firm table. the field can be null.
this is for the migration, sami found an occurrence of multiple referrals for the same patient, same service, same operation and with the openeyes schema there was no way to link the referral back to a firm or a consultant, only to a service.  adding this field in solves the problem for now.  when we create referrals in the system going forward this field will likely remain null, but we can use it to infer the correct firm for migrated referrals.