Resolves Jira issue OE-692
This was actually quite a scary bug. When you search for a patient and view the patient summary, it sets a yii session variable to the
id of the patient. This is then used as the id that events are created for, so if you searched for a patient and then browsed directly
to another patient via the diary, and clicked "add operation", it would have created the event for the first patient rather than the
one you were viewing. I've fixed this by setting the session variable when you view the episodes+events page, but i strongly question
the use of session variables at all, it seems a bit unnecessary and convoluted.