Resolves Jira issue OE-869
Red entries on the waiting list are not supposed to be printable as there is no due letter. Have disabled the print functionality for these items and made it pop up a box indicating that some items could not be printed and the reason why.
Have also used the same system for orange items where the patient has no GP.  In these cases we cannot print the GP letter as the patient has no GP.