Partially resolves jira issue OE-1278
Added a hidden flag to the event table.
Created a default scope in the event model to prevent returning any events where hidden = 1
Updated the gnarly manual queries for the diary and waiting list to ensure hidden events aren't returned there either.
One side effect of this change is that if you hide all the events for an episode, the episode still shows but with no events. Do we want to hide the episode in this case too?