OE-6265: Adds relationships to a subject
Adds a relationship for a subject to other subjects with a relationship type.
This included creating a generic patient lookup, a relationships view and
changes to the generic admin to allow actions to be taken after the save as
the system couldn't cope with creating many relationships where there was
and extra item in the table.