OE-6275: Implements adding subjects to studies
Adds the ability to propose a subject to a study on the subject page. The list
of studies is limited to those which the current user can propose which have
not ended.

Adds a list of studies which the subject was in but which have now ended.
The view for this has been added to the main views folder as this could
easily be used by any kind of study, genetic or otherwise.