OE-6276: Adds study participation to pivot.
Adds the status of the subjects participation in the study to the pivot table so
participation can be rejected.

Also refactors the previous studies template to be a generic studies list
template and removes check on user to drop down for study proposals
so it includes all studies as otherwise if you edit any study you do
not have permission to propose will be removed. Validation on submitting and
removing study proposals will be added to the model.