OE-6275: Validates studies to make sure the user can propose subject
Adds validation to check that any new studies added are validated to make sure
that the user proposing the subject is able to do so.

The trait was moved to the behaviour folder because it is basically performing
the same function. Considered changing over to a behaviour but as we currently
require none of the extra functionality it can stay as a trait for now.