Begins pedigree work
Creates the extra columns needed for a pedigree and versions the table. Also
starts work on the requirements for basic CRUD and search functions for the
pedigrees. As the existing code for this is terrible and does not work an complete
replacement of the existing functionality is being implemented.

OE-6268, OE-6267