Creates new genetic subject relationship
Because a patient may have many diagnoses which are not relevant to the
genetics pedigrees a new relationship is created between the genetics patient
and the disorder table. This will allow them to remove certain diagnosis that
are not required. Existing diagnosis are copied from all episodes and any
secondary diagnoses when a genetic patient is created.