Fixes problem in user edit with multiSelectList
In the user edit action for the auth roles the call to the multiSelectList
function had one more parameter passed than was defined, thus when I
created a new parameter for the through relationships it was being set as
a string and that caused problems in the template. After fixing this I then
checked all the other calls to make sure they weren't also sending extra
parameters and reformatted a lot of the templates at the same time.