Adds Internal Source selection
Adds the internal source selection to the database and creates a widget to select
elements to relate to it. The only issue here is that the table we are relating to is in
a different module so may or may not exist when the migration runs so this will need
to be checked.