OE-6340: Generates CSV and database log
This creates the CSV file and the database log. This required some refactoring
of the initial code as it wasn't checking for success on actually writing the files.
Basic checking has been added so the status will not be updated if the file
write fails and the logs won't be written unless the update is succesful.