OE-6198
Adds in duplication of Alpha Receptor Blocker to cover both eyes.
Deals with issue of adding a previously removed Eye side, and forcing of Diabetes and Alpha Receptor Blocker choice to that set on other side.
Inclusion of code that will remove the No or Not Known options if the patient had already had a diagnosis of Diabetes, only Yes is available on both eyes.