Fixes extracted_by
Extracted by was a complete mess, with some times it referring to a relationship
and others a simple text field. With extracted_by_id being used but never existing.
Renamed the extracted_by to extracted_by_text to store a text name on the import,
made a new column for extracted_by_id to store the actual user. Also removes the
is_technician from the user as it's not required in any way.