OE-7770 - Make inbox message previews replace line breaks with spaces
The inbox message previews on the homepage were previously ineffective
because they could often be very short due to line breaks. This commit
replaces those line breaks with whitespace characters to fit more into
the preview line.

To ensure that squeezed messages would always still be expandable (have
an expand icon), it was necessary to include two separate <div> elements
for each message in grid.php - one for the preview message and one for
the expanded message.

The changes in dashboard.php ensure that the expanded message starts off
hidden and that the expand icon shows/hides the preview and expanded
message <div> elements.

Also, isChangedByExpand is a new variable in isExpandableMessage() that
replaces isMultiLine for deciding when the Expand Icon is and is not
visible.