OE-7922 -Improve robustness by not using undocumented features of plotly
The fix in the previous commit was not robust. It relied on a feature
of plotly.js where it automatically assigns the 'number' data type to an
empty graph. This feature of plotly.js is undocumented, so it could
change in future updates without warning.

This commit provides a robust solution by avoiding this risk. It
explicitly checks if the axis type is 'date' before including the axis
in the calculation of the minimum and maximum dates. The axis type
'date' *is* a documented feature of plotly.js, so is much less likely to
break in future updates.