OE-7996 - Make 'Add Event' button work on Message View page
On closer inspection, I realised that this bug was caused by a
javascript error. This error was caused by module.js - the js was
attempting to add an Autocomplete dialog to an input field, even when
the input field in question was not on this page.

This commit places this javascript inside a conditional block that
checks whether the input field with id 'oe-autocompletesearch' is in the
DOM before attempting to initialize an Autocomplete dialog for it.