Added a check for undefined diagnoses for Download CSV button
This bug will only appear if If there is a patient without a diagnoses. this bug is caused by the Javascript not handling the undefined value for the diagnosis, when it attempts to replace "," with "|" to prevent issue with the CSV output file.

This fix is to add a check to see if the value is undefined, if so return an empty string for the CSV.