OE-8946 Fixed management summary history view.
These views attempted to read data directly from the model but the
available attributes have changed since it was written. Comments were
still available under a different name but only user id is now available
not username. So switched the method of getting the data to the one used
in other views, using the exam_api.