OE-8700 Fixed dose unit not shown on validation error.
Entire form is set up to hide form elements until clicked on, after
which the inputs are shown. Changed to disable hiding when there are
validation errors, also disabled hiding when no dose is input at all as
when this is done, there is no way to add a dose later when editing.