OE-9254 Fixed OEScape crash.
In commit 6a540485fe4b70f64bd33e0df32f4a3ab12d5fe0 the `run_oescape`
function was given an argument in the Visual Actuity History widget of
which the Medical Retinal History widget is a child. the MRH method does
not specify any arguments, php5 is allowing this overload but php7 is
not.