OE-9539 Fixed bug with dose unit term
When legacy dose is of the form 124mg bd for eg. after splitting legacy
dose has a count of 2, hence the preg_match was added to account for
this. However the second part of the unit still needs to be added on to
the end of the dose unit in this case.