OE-9631 Make regex more strict.
This will remove all tags except those that are allowed, side effect is
that it will also remove anything that happens to be between angle
braces. For example, if on one line there is "x<3 and x>1" the middle
part will be removed. However there is no way to avoid this without
trying to maintain a blacklist.