OE-9906 Fix ambiguous letter reports.
Could not replicate issue exactly as laid out in ticket, instead when
a correspondence is 'Save and Print'-ed the report shows two entries for
the letter in question. One is for the docman output and one is for the
print output. This is because both print and electronic were selected,
if only one is selected then there is only one entry in the report. To
fix the issue of these two entries being ambiguous (as they are
identical in the report), the output type is appended to the 'Type'
field of the report.