Feature/oe 11855 (#6702)
* Added data model for discharge status and destination.
Incorporated new discharge data model into follow-up element as selectable options if the selected status is Discharge (this can be customised to apply to other statuses as well).
Added institution selection if the selected discharge destination is 'Transfer to another facility'. In a future commit this will change to an autocomplete search field. This can be customised to apply to other discharge destinations as well.
Added JS validation for new discharge options to ensure that a discharge destination and status are selected, and to ensure an institution is selected if the selected destination is 'Transfer to another facility'.

* Added admin screens for managing discharge statuses and destinations.
Moved 'select' check boxes code to only display a checkbox if the model has a display_order column (maintains consistency with how the checkall checkbox is displayed).
Added 'Show Discharge Options' checkbox to the Follow-up Statuses admin screen.

* Institution transfer field in adder dialog for discharge options now uses a search field to populate its list.

* Examination view now shows the full follow-up string for the discharge status.
Fixed typo in hidden form field for transfer_institution_id.

Co-authored-by: Grant Morphett <Grant.Morphett@ToukanLabs.com>
Co-authored-by: Peichun Zhou <peichun.zhou@toukanlabs.com>